Handle switching between different installation of Transmission
When switching between different installation of Transmission,
the base_id of all tasks are completely different and any
old tasks that have existed cannot be reused. Set the base_id
of tasks that no longer exist to -1 so that we can avoid old
data tainting the behaviour after the switch.

Change-Id: Id90052c9c3c3edc7f196c37b12db793094c19333